Dropdown Fixes
A majority of the dropdown feature working. Left Shoulder waiting on new
animations. AnimationController moved to PatientController.
