+++ Assets/Models.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f1e9c0e09bac66f449b4c4f78c112223
+folderAsset: yes
+timeCreated: 1481665248
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
