--- Assets/Scripts/AnimationController.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 6a73f811b12d27a4cbd21cc8d1bffc8a
-timeCreated: 1480908101
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
