+++ Library/UnityAssemblies/UnityEngine.Analytics.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8" standalone="yes"?>
+<doc>
+  <members>
+    <assembly>
+      <name>UnityEngine.Analytics</name>
+    </assembly>
+    <member name="P:UnityEngine.Analytics.AnalyticsTracker.eventName">
+      <summary>
+        <para>Name of the event.</para>
+      </summary>
+    </member>
+    <member name="M:UnityEngine.Analytics.AnalyticsTracker.TriggerEvent">
+      <summary>
+        <para>Trigger the instrumented event.</para>
+      </summary>
+    </member>
+  </members>
+</doc>
