+++ Library/UnityAssemblies/UnityEngine.PlaymodeTestsRunner.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8" standalone="yes"?>
+<doc>
+  <members>
+    <assembly>
+      <name>UnityEngine.PlaymodeTestsRunner</name>
+    </assembly>
+  </members>
+</doc>
