+++ Library/UnityAssemblies/UnityEngine.VR.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8" standalone="yes"?>
+<doc>
+  <members>
+    <assembly>
+      <name>UnityEngine.VR</name>
+    </assembly>
+  </members>
+</doc>
